backend_django/requirements.txt
# Django Backend for Indian Cattle & Buffalo Breed Recognition

## Requirements
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-spectacular>=0.27.0
whitenoise>=6.6.0
Pillow>=10.0.0

# ML Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Optional ML Dependencies
pytorch-grad-cam>=1.5.0

# Database (optional - for production)
psycopg2-binary>=2.9.9

# Caching (optional)
redis>=5.0.0
django-redis>=5.4.0

# Async task queue (optional)
celery>=5.3.0

# Development
python-dotenv>=1.0.0
